chore : code optimized
