+++ insights/urls.py
@@ -1,11 +1,9 @@
 from django.urls import path
 from . import views
 
-app_name = "vault"
+app_name = "insights"
 
 urlpatterns = [
-    path("", views.vault_home, name="home"),
+    path("", views.insights_home, name="home"),
     # 🔥 category + slug
-    path("<str:category>/<slug:slug>", views.vault_detail, name="detail"),
-    # path("vault/<slug:slug>/", views.vault_detail, name="detail"),
-]
+    path("<str:category>/<slug:slug>", views.insights_home_detail, name="detail"),]
