+++ insights/views.py
@@ -1,8 +1,8 @@
-from django.http import HttpResponse
+
 from django.shortcuts import render, get_object_or_404
 from .models import InsightsPost
 
-def vault_home(request):
+def insights_home(request):
     posts = InsightsPost.objects.filter(is_published=True)
 
     context = {
@@ -10,20 +10,20 @@ def vault_home(request):
         "blogs": posts.filter(category="blog"),
         "offers": posts.filter(category="offer"),
     }
-    return render(request, "vault/vault_home.html", context)
+    return render(request, "insights/insights_home.html", context)
 
 
-def vault_detail(request,category, slug):
+def insights_home_detail(request,category, slug):
     post = get_object_or_404(InsightsPost,slug=slug,category=category,is_published=True)
 
     response = render(
         request,
-        "vault/vault_detail.html",
+        "insights/insights_detail.html",
         {"post": post}
     )
     return response
 
-# def vault_detail(request, category, slug):
+# def insights_detail(request, category, slug):
 #     # Only raise error for testing
 #     res = 1 / 0  # force exception
 #     return HttpResponse(res)
