+++ payments/urls.py
@@ -11,4 +11,5 @@ urlpatterns = [
     path("failed/", views.payment_failed, name="payment_failed"),
     path("retry/<str:order_id>/", views.retry_payment, name="retry_payment"),
     path("webhook/razorpay/", views.razorpay_webhook, name="razorpay_webhook"),
+    path('<int:pk>/payment-result/', views.payment_result, name='payment_result'),
 ]
