+++ products/views.py
@@ -110,33 +110,3 @@ def download_file(request, pk):
     except Exception as e:
         messages.error(request, f"Download failed: {str(e)}")
         return redirect('products:detail', pk=pk)
-
-# ======================
-# Payment Result Page
-# ======================
-def payment_result(request, pk):
-    session_key = f"paid_{pk}"
-    product = get_object_or_404(Product, pk=pk, is_active=True)
-
-    # 1. Check kara product Free aahe ka
-    is_free = product.price == 0
-
-    # 2. Database check (Paid products sathi)
-    db_paid = Payment.objects.filter(product=product, status="SUCCESS", paid=True).exists()
-
-    # 3. Success condition: Jar Free asel OR Session madhe entry asel OR DB madhe entry asel
-    if is_free or request.session.get(session_key) or db_paid:
-        status = "success"
-        file_url = reverse("products:download_file", args=[pk])
-        # Download access sathi session set kara
-        request.session[session_key] = True 
-    else:
-        status = "failed"
-        file_url = None
-
-    return render(request, "products/payment_result.html", {
-        "status": status,
-        "product": product,
-        "file_url": file_url,
-        "is_free": product.price == 0, # <--- He add kara
-    })
